payment_processor/requirements.txt
aiokafka==0.8.0
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
fastapi==0.92.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
kafka-python==2.0.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pydantic==1.10.5
python-dateutil==2.8.2
pytz==2022.7.1
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
